--- lib/before_file/data/domain/model/conversation.dart
@@ -1,12 +1,0 @@
-// Temp Model
-class Conversation {
-  final String username;
-  final String message;
-  final bool isMe;
-
-  const Conversation({
-    required this.username,
-    required this.message,
-    required this.isMe,
-  });
-}
