--- lib/before_file/home_and_talk/home.dart
@@ -1,180 +1,0 @@
-import 'package:flutter/material.dart';
-import 'more_info.dart';
-import 'profile_screen.dart';
-import 'talk.dart';
-
-class Home extends StatefulWidget {
-  const Home({super.key});
-
-  @override
-  State<Home> createState() => _HomeState();
-}
-
-class _HomeState extends State<Home> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        decoration: const BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.bottomLeft,
-            end: Alignment.topRight,
-            colors: [
-              Colors.black,
-              Colors.lightBlue,
-            ],
-          ),
-        ),
-        child: Column(
-          children: [
-            const SizedBox(
-              height: 75,
-            ),
-            Container(
-              margin: const EdgeInsets.all(50),
-              decoration: BoxDecoration(
-                color: Colors.black87,
-                borderRadius: BorderRadius.circular(
-                  20,
-                ),
-              ),
-              child: const Padding(
-                padding: EdgeInsets.all(75),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Column(
-                      children: [
-                        Text(
-                          '안녕하세요',
-                          style: TextStyle(
-                            fontSize: 32,
-                            fontWeight: FontWeight.w600,
-                            color: Colors.white,
-                          ),
-                        ),
-                        SizedBox(
-                          height: 20,
-                        ),
-                        Row(
-                          children: [
-                            Text(
-                              '(사용자) 님',
-                              style: TextStyle(
-                                fontSize: 30,
-                                color: Colors.white,
-                              ),
-                            ),
-                          ],
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-              ),
-            ),
-            Container(
-              width: 250,
-              height: 250,
-              decoration: const BoxDecoration(
-                color: Color(0xFFECE6CC),
-                shape: BoxShape.circle,
-              ),
-              child: TextButton(
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => const Talk(),
-                    ),
-                  );
-                },
-                child: const Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text(
-                      '대화 시작하기',
-                      style: TextStyle(
-                        fontSize: 32,
-                        fontWeight: FontWeight.w600,
-                        color: Colors.black,
-                      ),
-                    ),
-                    SizedBox(
-                      height: 5,
-                    ),
-                    Icon(
-                      Icons.mic,
-                      size: 50,
-                      color: Colors.deepPurpleAccent,
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            const Spacer(),
-            SizedBox(
-              height: 50,
-              width: double.infinity,
-              child: DecoratedBox(
-                decoration: const BoxDecoration(
-                  color: Colors.grey,
-                ),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: [
-                    IconButton(
-                      onPressed: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                            builder: (context) => const Home(),
-                          ),
-                        );
-                      },
-                      icon: const Icon(
-                        Icons.home,
-                        size: 40,
-                        color: Colors.deepPurpleAccent,
-                      ),
-                    ),
-                    IconButton(
-                      onPressed: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                            builder: (context) => const ProfileScreen(),
-                          ),
-                        );
-                      },
-                      icon: const Icon(
-                        Icons.person,
-                        size: 40,
-                        color: Colors.deepPurpleAccent,
-                      ),
-                    ),
-                    IconButton(
-                      onPressed: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                            builder: (context) => const MoreInfo(),
-                          ),
-                        );
-                      },
-                      icon: const Icon(
-                        Icons.more_horiz,
-                        size: 40,
-                        color: Colors.deepPurpleAccent,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
