--- lib/before_file/home_and_talk/login.dart
@@ -1,78 +1,0 @@
-import 'package:flutter/material.dart';
-import 'home.dart';
-
-class Login extends StatefulWidget {
-  const Login({super.key});
-
-  @override
-  State<Login> createState() => _LoginState();
-}
-
-class _LoginState extends State<Login> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        decoration: const BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.bottomLeft,
-            end: Alignment.topRight,
-            colors: [
-              Colors.black,
-              Colors.lightBlue,
-            ],
-          ),
-        ),
-        child: Column(
-          children: [
-            const SizedBox(
-              height: 75,
-            ),
-            const Image(
-              image: AssetImage('assets/logo.png'),
-            ),
-            Container(
-              width: 250,
-              height: 250,
-              decoration: const BoxDecoration(
-                color: Color(0xFFECE6CC),
-                shape: BoxShape.circle,
-              ),
-              child: TextButton(
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => const Home(),
-                    ),
-                  );
-                },
-                child: const Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text(
-                      'Google 로그인',
-                      style: TextStyle(
-                        fontSize: 32,
-                        fontWeight: FontWeight.w600,
-                        color: Colors.black,
-                      ),
-                    ),
-                    SizedBox(
-                      height: 5,
-                    ),
-                    Icon(
-                      Icons.login,
-                      size: 50,
-                      color: Colors.deepPurpleAccent,
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
