--- lib/before_file/home_and_talk/profile_screen.dart
@@ -1,86 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'home.dart';
-import 'more_info.dart';
-
-class ProfileScreen extends StatefulWidget {
-  const ProfileScreen({super.key});
-
-  @override
-  State<ProfileScreen> createState() => _ProfileScreenState();
-}
-
-class _ProfileScreenState extends State<ProfileScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('프로필 화면'),
-      ),
-      body: Column(
-        children: [
-          const Spacer(),
-          SizedBox(
-            height: 50,
-            width: double.infinity,
-            child: DecoratedBox(
-              decoration: const BoxDecoration(
-                color: Colors.grey,
-              ),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  IconButton(
-                    onPressed: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => const Home(),
-                        ),
-                      );
-                    },
-                    icon: const Icon(
-                      Icons.home,
-                      size: 40,
-                      color: Colors.deepPurpleAccent,
-                    ),
-                  ),
-                  IconButton(
-                    onPressed: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => const ProfileScreen(),
-                        ),
-                      );
-                    },
-                    icon: const Icon(
-                      Icons.person,
-                      size: 40,
-                      color: Colors.deepPurpleAccent,
-                    ),
-                  ),
-                  IconButton(
-                    onPressed: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => const MoreInfo(),
-                        ),
-                      );
-                    },
-                    icon: const Icon(
-                      Icons.more_horiz,
-                      size: 40,
-                      color: Colors.deepPurpleAccent,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
