--- lib/before_file/presentation/feedback_screen.dart
@@ -1,31 +1,0 @@
-import 'package:flutter/material.dart';
-
-class FeedbackScreen extends StatelessWidget {
-  const FeedbackScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Feedback'),
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back),
-          onPressed: () {
-            Navigator.pop(context); // This will pop the FeedbackScreen
-          },
-        ),
-      ),
-      body: const Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Center(
-            child: Text(
-              '피드백입니다.',
-              style: TextStyle(fontSize: 24),
-            ),
-          ),
-        ],
-      ), // Rest of your FeedbackScreen content here
-    );
-  }
-}
