+++ lib/domain/ai_chat/constants/themes.dart
