+++ lib/domain/ai_chat/models/chat_model.dart
