+++ lib/domain/ai_chat/providers/active_theme_provider.dart
