+++ lib/domain/ai_chat/providers/chats_provider.dart
