+++ lib/domain/ai_chat/services/ai_handler.dart
