+++ lib/domain/ai_chat/widgets/theme_switch.dart
