+++ lib/domain/ai_chat/widgets/toggle_button.dart
