--- lib/presentation/screen/before_talk_screen.dart
@@ -1,1 +1,0 @@
-
