--- lib/presentation/screen/my_page_screen.dart
@@ -1,1 +1,0 @@
-
