--- lib/presentation/screen/pay_reset_screen.dart
@@ -1,1 +1,0 @@
-
