--- lib/presentation/screen/z_speak_screen.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SpeakScreen extends StatefulWidget {
-  const SpeakScreen({super.key});
-
-  @override
-  State<SpeakScreen> createState() => _SpeakScreenState();
-}
-
-class _SpeakScreenState extends State<SpeakScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text('data'),
-      ),
-    );
-  }
-}
