+++ pubspec.yaml
@@ -45,12 +45,9 @@ dependencies:
   firebase_storage: ^11.7.4
   cloud_firestore: ^4.17.2
   google_sign_in: ^6.2.1
-  flutter_dotenv: ^5.1.0
   envied: ^0.5.4+1
-  googleapis: ^13.1.0
   intl: ^0.19.0
   flutter_riverpod: ^2.5.1
-  chat_gpt_sdk: ^3.0.7
   speech_to_text: ^6.6.1
   flutter_tts: ^3.8.1
   dart_openai: ^5.1.0
